\name{maxLinear-internal}
\title{Internal maxLinear objects}
\alias{.HHJs.sampleHits}
\alias{.MHG.sampleHits}
\alias{.MHG.weights}
\alias{.readHHJs}
\alias{.readMHG}
\alias{.sampleHits}
\description{Internal maxLinear objects.}
\details{These are not to be called by the user.}
\keyword{internal}